Addiing Audit Log Functionality and updated file structure
